The PSA import functionality has been moved to a file import dialog.
As a result, the "PSA Import" panel in the Armature Data tab has been
removed as it is now redundant.

This was made possible by https://developer.blender.org/D15543.

As a result, the minimum Blender version has now been bumped to 3.4.

The 4.2.0 version is now in LTS mode and will not be receiving new
features.