Fixed a bug where the importer would crash when importing PSA files from programs using CUE4Parse (#46)
This is not actually a bug in this addon, but a bug with the CUE4Parse library itself. A fix for the library has been submitted (https://github.com/FabianFG/CUE4Parse/issues/103), but there will still be bugged versions hanging around for a while, so it's prudent to have this addon accomodate the broken files, especially since the fix is relatively easy and won't interfere with properly formatted files.

At some point, this fix will be deprecated to reduce cruft.