PSA sequences are now selected by default
There were multiple bug reports from users who were getting mixed
signals from the addon, believing that the bone name warnings were
the cause of the sequences not being imported. The actual issue was that
users didn't know they needed to manually select the sequences within
the PSA.

This should fix the poor UX around this, as just having it selected by
default is more sensible for a wider audience.