Fix #137: Shape keys start with 1.0 value on PSK import
There was an undocumented change in 5.0 where the shape keys now start with a weight of 1.0 instead of 0.0 when created.

I have added an assert in the shape key test to make sure that the values are set to `0.0`.

This also fixes the assertion in the poll function (asserts should never be in poll functions).